Added automatic namespacing to the markdown parser by passing a different base_url.
A link like "[Message](files/messages)" will be given the link "guide/kohana/files/messages".
This means that linking to a different module is essentially broken, unless you do an absolute link.
The solutions are to make all links in the menu.md and links between pages in the same module have the module name in front of them, or add a special syntax for linking to a different module.