Hide class prefixes other than Kohana_ in userguide - implements #3529
Userguide now uses Kodoc::is_transparent() method to determine whether
a class is a transparent extension class or not, allowing developers to
use their own prefixes for classes and keep them out of the kohana core
namespace, without losing the ability to hide the implementing classes in
the userguide.

Previously only classes beginning Kohana_ were hidden, but the prefixes
can now be configured.